metadjang/requirements.txt
celery==3.1.20
Django==1.10.5
django-dotenv==1.4.1
django-jet==1.0.4
django-webpack-loader==0.4.1
djangorestframework==3.5.4
psycopg2==2.6.2
redis==2.10.5
wheel==0.29.0